+++ src/routes/v1/room.routes.js
@@ -6,9 +6,9 @@ import {
     updateRoom,
     addUserToRoom,
     disconnectUserFromRoom,
-    acceptOrRejectPendingUser
+    acceptOrRejectPendingUser,
+    announce
 } from '../../controllers/room.controllers.js';
-import { announce } from '../../controllers/announce.controllers.js';
 import { authVerify } from '../../middlewares/userAuth.middleware.js';
 import prisma from '../../config/db.config.js';
 
@@ -20,6 +20,7 @@ roomRouter.patch('/update', authVerify, updateRoom);
 roomRouter.post('/user/add', authVerify, addUserToRoom);
 roomRouter.post('/user/disconnect', authVerify, disconnectUserFromRoom);
 roomRouter.post('/user/pending', authVerify, acceptOrRejectPendingUser);
+roomRouter.post("/announce", authVerify, announce);
 
 // just  testing routes for now can be upgraded later
 roomRouter.get('/', async (req, res) => {
@@ -36,8 +37,6 @@ roomRouter.delete('/delete/:code', async (req, res) => {
     });
     return res.status(200).send('Room deleted');
 });
-
-roomRouter.post("/announce", announce);
 
 export default roomRouter;
